API documentation

API documentation

Index

FEMTruss

FEMTruss.Truss

FEMBase.assemble!

FEMTruss.get_Kg_and_Tg

FEMTruss — Module.

Truss implementation for JuliaFEM.

source

FEMTruss.Truss — Type.

This truss fromulation is from Concepts and applications of finite element analysis, forth edition Chapter 2.4 Cook, Malkus, Plesha, Witt

Features

Nodal forces can be set using element Poi1 with field nodal force i, where i is dof number.

Displacements can be fixed using element Poi1 with field fixed displacement i, where i is dof number.

source

FEMTruss.get_Kg_and_Tg — Method.

get_Kg_and_Tg(element::Element{Seg2}, nnodes, ndim, time) This function assembles the local stiffness uses global transformation matrix to make the global version of the local stiffnes matrix Need these to find element forces later Will need to change allocation strategy to pre-allocation later Can discuss if we really need nnodes, since that is always 2 for trusses

source

FEMBase.assemble! — Method.

assemble!(assembly:Assembly, problem::Problem{Elasticity}, elements, time)

Start finite element assembly procedure for Elasticity problem. Function groups elements to arrays by their type and assembles one element type at time. This makes it possible to pre-allocate matrices common to same type of elements.

source